pyproject.toml
[tool.poetry]
name = "lf-web"
version = "0.1.0"
description = "Website for Lejre Fitness"
authors = ["Carsten Skov <carsten@simcax.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "2.3.3"
Flask-Session = "^0.5.0"
gunicorn = "^21.2.0"
Jinja2 = "^3.1.3"
redis = "^5.0.1"
loguru = "^0.7.2"
werkzeug = "2.3.8"
foreninglet-data = ">=0.3.3"
markdown = "^3.5.2"
markupsafe = "^2.1.5"
sentry-sdk = "^1.41.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
sqlalchemy-utils = "^0.41.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
docker = "^7.0.0"
pytest = "^7.4.4"
testcontainers-redis = "^0.0.1rc1"
pylint = "^3.0.3"
pytest-cov = "^4.1.0"
testcontainers-postgres = "^0.0.1rc1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
